% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_mt_susie}
\alias{fit_mt_susie}
\title{Estimate a multi-trait Sum of Single Effect regression model}
\usage{
fit_mt_susie(
  x,
  y,
  levels = 15L,
  max_iter = 100L,
  min_iter = 1L,
  coverage = 0.9,
  tol = 1e-08,
  prior_var = 0.01,
  lodds_cutoff = 0,
  min_pip_cutoff = NULL,
  full_stat = TRUE
)
}
\arguments{
\item{x}{design matrix}

\item{y}{output matrix}

\item{levels}{number of "single" effects}

\item{max_iter}{maximum iterations}

\item{min_iter}{minimum iterations}

\item{tol}{tolerance}

\item{prior_var}{prior variance}

\item{lodds_cutoff}{log-odds cutoff in trait selection steps}

\item{min_pip_cutoff}{minimum PIP cutoff in building credible sets}

\item{full_stat}{keep full statistics}
}
\value{
a list of mtSusie results
\item{alpha}{Posterior probability of variant across \verb{p x level}}
\item{resid.var}{residual variance \verb{level x traits}}
\item{prior.var}{prior variance \verb{1 x traits}}
\item{log.odds}{log-odds ratio \verb{level x traits}}
\item{mu}{a list of \verb{p x m} mean parameters}
\item{var}{a list of \verb{p x m} variance parameters}
\item{lbf}{a list of \verb{p x m} log-Bayes Factors}
\item{z}{a list of \verb{p x m} z-scores}
\item{n}{number of samples}
\item{m}{number of traits/outputs}
\item{p}{number of variants}
\item{L}{number of layers/levels}

\item{cs}{credible sets (use \code{data.table::setDT} to assemble)}

In the \code{cs}, we have
\item{variants}{variant indexes \verb{[1 .. p]}}
\item{traits}{trait indexes \verb{[1 .. m]}}
\item{levels}{levels \verb{[1 .. L]}}
\item{alpha}{shared PIP}
\item{mean}{univariate mean}
\item{var}{univariate var}
\item{z}{univariate z-score}
\item{lfsr}{local false sign rate}
\item{lodds}{log-odds ratio}
}
\description{
Estimate a multi-trait Sum of Single Effect regression model
}
