% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtSusie.R
\name{mt_susie}
\alias{mt_susie}
\title{Fit multi-trait SuSiE across several regression models}
\usage{
mt_susie(
  X,
  Y,
  L = 15,
  clamp = 4,
  max.iter = 100,
  min.iter = 5,
  tol = 1e-08,
  coverage = 0.9,
  min.pip.cutoff = NULL,
  lodds.cutoff = 0,
  prior.var = 100
)
}
\arguments{
\item{X}{n x p design matrix}

\item{Y}{n x m output matrix}

\item{L}{levels or expected num of independent factors (default: 15)}

\item{clamp}{handle outliers by winsorization (default: 4)}

\item{max.iter}{maximum iterations (default: 100)}

\item{min.iter}{minimum iterations (default: 5)}

\item{tol}{tolerance level check convergence (default: 1e-8)}

\item{coverage}{targeted PIP coverage per each level (default: 0.9)}

\item{min.pip.cutoff}{min PIP to report (default: 1/p)}

\item{lodds.cutoff}{log-odds ratio cutoff for multitrait factors}

\item{prior.var}{prior level variance}
}
\description{
Fit multi-trait SuSiE across several regression models
}
