% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_mt_interaction_susie}
\alias{fit_mt_interaction_susie}
\title{Estimate mtSusie regression model with an interaction matrix}
\usage{
fit_mt_interaction_susie(
  x,
  y,
  w,
  levels_per_wx = 3L,
  levels_per_w = 1L,
  levels_per_x = 1L,
  max_iter = 100L,
  min_iter = 1L,
  coverage = 0.9,
  tol = 1e-04,
  prior_var = 0.1,
  min_pip_cutoff = NULL,
  full_stat = TRUE,
  update_prior = FALSE,
  do_hard_selection = FALSE,
  hard_lodds_cutoff = 0
)
}
\arguments{
\item{x}{design matrix}

\item{y}{output matrix}

\item{w}{interaction matrix}

\item{levels_per_wx}{number of "single" effects for w*x terms}

\item{levels_per_w}{number of "single" effects for w terms}

\item{levels_per_x}{number of "single" effects for x terms}

\item{max_iter}{maximum iterations}

\item{min_iter}{minimum iterations}

\item{tol}{tolerance}

\item{prior_var}{prior variance}

\item{min_pip_cutoff}{minimum PIP cutoff in building credible sets}

\item{full_stat}{keep full statistics}

\item{update_prior}{update prior variance or not}
}
\value{
a list of mtSusie results
\item{alpha}{Posterior probability of variant across \verb{p x level}}
\item{resid.var}{residual variance \verb{level x traits}}
\item{prior.var}{prior variance \verb{1 x traits}}
\item{log.odds}{log-odds ratio \verb{level x traits}}
\item{mu}{a list of \verb{p x m} mean parameters}
\item{var}{a list of \verb{p x m} variance parameters}
\item{lbf}{a list of \verb{p x m} log-Bayes Factors}
\item{z}{a list of \verb{p x m} z-scores}
\item{n}{number of samples}
\item{m}{number of traits/outputs}
\item{p}{number of variants}
\item{L}{number of layers/levels}

\item{cs}{credible sets (use \code{data.table::setDT} to assemble)}

In the \code{cs}, we have
\item{variant}{variant indexes \verb{[1 .. p]}}
\item{trait}{trait indexes \verb{[1 .. m]}}
\item{level}{levels \verb{[1 .. L]}}
\item{alpha}{shared PIP}
\item{mean}{univariate mean}
\item{var}{univariate var}
\item{z}{univariate z-score}
\item{lfsr}{local false sign rate}
\item{lodds}{log-odds ratio}
\item{interaction}{interacting column index}
}
\description{
Estimate mtSusie regression model with an interaction matrix
}
