% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_mt_susie}
\alias{fit_mt_susie}
\title{Estimate a multi-trait Sum of Single Effect regression model}
\usage{
fit_mt_susie(
  x,
  y,
  levels = 15L,
  max_iter = 100L,
  min_iter = 5L,
  coverage = 0.9,
  tol = 1e-08,
  prior_var = 100,
  lodds_cutoff = 0,
  min_pip_cutoff = NULL
)
}
\arguments{
\item{x}{design matrix}

\item{y}{output matrix}

\item{levels}{number of "single" effects}

\item{max_iter}{maximum iterations}

\item{min_iter}{minimum iterations}

\item{tol}{tolerance}

\item{prior_var}{prior variance}
}
\value{
a list of mtSusie results

\item{alpha}{Posterior probability of variant across k traits; \code{alpha[j,l]}}
\item{resid.var}{Residual variance; \code{alpha[l,k]}}
}
\description{
Estimate a multi-trait Sum of Single Effect regression model
}
