% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtSusie.R
\name{mt_susie}
\alias{mt_susie}
\title{Fit multi-trait SuSiE across several regression models}
\usage{
mt_susie(
  X,
  Y,
  L = 5,
  max.iter = 100,
  min.iter = 5,
  tol = 1e-08,
  coverage = 0.9,
  min.pip.cutoff = NULL,
  prior.var = 0.01,
  output.full.stat = FALSE,
  local.residual = FALSE,
  update.prior = TRUE
)
}
\arguments{
\item{X}{n x p design matrix}

\item{Y}{n x m output matrix}

\item{L}{levels or expected num of independent factors (default: 5)}

\item{max.iter}{maximum iterations (default: 100)}

\item{min.iter}{minimum iterations (default: 5)}

\item{tol}{tolerance level check convergence (default: 1e-8)}

\item{coverage}{targeted PIP coverage per each level (default: 0.9)}

\item{min.pip.cutoff}{min PIP to report (default: 1/p)}

\item{prior.var}{prior level variance (default: 1e-2)}

\item{output.full.stat}{output every bit of the results (default: FALSE)}

\item{local.residual}{calculate residual variance locally (default: FALSE)}

\item{update.prior}{update prior variance (default: TRUE)}
}
\value{
a list of mtSusie estimates

\item{alpha}{Posterior probability of variant across \verb{p x level}}
\item{resid.var}{residual variance \verb{level x traits}}
\item{prior.var}{prior variance \verb{1 x traits}}
\item{log.odds}{log-odds ratio \verb{level x traits}}
\item{mu}{a list of \verb{p x m} mean parameters}
\item{var}{a list of \verb{p x m} variance parameters}
\item{lbf}{a list of \verb{p x m} log-Bayes Factors}
\item{z}{a list of \verb{p x m} z-scores}
\item{n}{number of samples}
\item{m}{number of traits/outputs}
\item{p}{number of variants}
\item{L}{number of layers/levels}
\item{loglik}{log-likelihood trace}
\item{cs}{credible sets (use \code{data.table::setDT} to assemble)}

In the \code{cs}, we have
\item{variant}{variant indexes \verb{[1 .. p]}}
\item{trait}{trait indexes \verb{[1 .. m]}}
\item{level}{levels \verb{[1 .. L]}}
\item{alpha}{shared PIP}
\item{mean}{univariate mean}
\item{var}{univariate var}
\item{z}{univariate z-score}
\item{lfsr}{local false sign rate}
\item{lodds}{log-odds ratio}
}
\description{
Fit multi-trait SuSiE across several regression models
}
